API de correos estudiantes y propietarios

+++ src/main/java/com/ViviEstu/repository/DatosPropiedadesRepository.java
@@ -3,7 +3,7 @@ package com.ViviEstu.repository;
 import com.ViviEstu.model.entity.DatosPropiedades;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DatosPropiedadesRepository extends JpaRepository<DatosPropiedades, Long> {
+public interface DatosPropiedadesRepository extends JpaRepository<DatosPropiedades, Integer> {
 
     boolean existsByDniPropietarioAndNroPartida(String dni, String nroPartida);
 }
