+++ src/main/java/com/ViviEstu/repository/DatosUniversitariosRepository.java
@@ -3,7 +3,7 @@ package com.ViviEstu.repository;
 import com.ViviEstu.model.entity.DatosUniversitarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DatosUniversitariosRepository extends JpaRepository<DatosUniversitarios, Long>
+public interface DatosUniversitariosRepository extends JpaRepository<DatosUniversitarios, Integer>
 {
     boolean existsByCorreoInstitucional(String correoInstitucional);
